requirements.txt
appdirs==1.4.4
attrs==21.4.0
cattrs==1.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cycler==0.11.0
Deprecated==1.2.13
distlib==0.3.4
filelock==3.6.0
Flask==2.1.1
Flask-Cors==3.0.10
fonttools==4.33.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
platformdirs==2.5.2
plotly==5.7.0
pycparser==2.21
PyGithub==1.55
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
requests-cache==0.9.4
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
tenacity==8.0.1
tinydb==4.7.0
url-normalize==1.4.3
urllib3==1.26.9
virtualenv==20.14.1
Werkzeug==2.1.1
wrapt==1.14.0
